[GN-48] rename: provider layout 파일 이동
