--- src/components/QueryProvider/QueryProvider.tsx
@@ -1,21 +1,0 @@
-"use client";
-
-import { QueryClient, QueryClientProvider } from "@tanstack/react-query";
-import { ReactQueryDevtools } from "@tanstack/react-query-devtools";
-import { ReactNode, useState } from "react";
-
-const createQueryClient = () =>
-  new QueryClient({
-    defaultOptions: {},
-  });
-
-export default function QueryProvider({ children }: { children: ReactNode }) {
-  const [client] = useState(() => createQueryClient());
-
-  return (
-    <QueryClientProvider client={client}>
-      {children}
-      <ReactQueryDevtools initialIsOpen />
-    </QueryClientProvider>
-  );
-}
